refactor: 회원 상태, 권한 enum
