--- src/main/java/shop/nuribooks/books/member/member/entity/AuthorityEnum.java
@@ -1,13 +1,0 @@
-package shop.nuribooks.books.member.member.entity;
-
-import com.fasterxml.jackson.annotation.JsonValue;
-
-public enum AuthorityEnum {
-
-	ADMIN, MEMBER, SELLER;
-
-	@JsonValue
-	public String getValue() {
-		return name();
-	}
-}
