--- src/main/java/shop/nuribooks/books/member/member/entity/StatusEnum.java
@@ -1,6 +1,0 @@
-package shop.nuribooks.books.member.member.entity;
-
-public enum StatusEnum {
-
-	ACTIVE, INACTIVE, WITHDRAWN
-}
