+++ src/test/java/shop/nuribooks/books/member/address/service/AddressServiceTest.java
@@ -2,8 +2,7 @@ package shop.nuribooks.books.member.address.service;
 
 import static java.math.BigDecimal.ZERO;
 import static org.assertj.core.api.Assertions.*;
-import static shop.nuribooks.books.member.member.entity.AuthorityEnum.MEMBER;
-import static shop.nuribooks.books.member.grade.entity.GradeEnum.STANDARD;
+import static shop.nuribooks.books.member.member.entity.AuthorityType.MEMBER;
 
 import java.math.BigDecimal;
 import java.time.LocalDate;
@@ -23,7 +22,7 @@ import shop.nuribooks.books.member.customer.entity.Customer;
 import shop.nuribooks.books.member.grade.entity.Grade;
 import shop.nuribooks.books.member.grade.repository.GradeRepository;
 import shop.nuribooks.books.member.member.entity.Member;
-import shop.nuribooks.books.member.member.entity.StatusEnum;
+import shop.nuribooks.books.member.member.entity.StatusType;
 import shop.nuribooks.books.exception.address.AddressNotFoundException;
 import shop.nuribooks.books.member.address.repository.AddressRepository;
 import shop.nuribooks.books.member.customer.repository.CustomerRepository;
@@ -173,7 +172,7 @@ class AddressServiceTest {
                 .authority(MEMBER)
                 .grade(grade)
                 .userId("nuriaaaaaa")
-                .status(StatusEnum.ACTIVE)
+                .status(StatusType.ACTIVE)
                 .birthday(LocalDate.of(1988, 8, 12))
                 .createdAt(LocalDateTime.now())
                 .point(ZERO)
